Only prioritise a task if there's time to run it

+++ src/main/osd/osd.c
@@ -1171,8 +1171,8 @@ void osdUpdate(timeUs_t currentTimeUs)
 {
     // Times are known to be short so use uint16_t rather than timeUs_t
     static uint16_t osdStateDurationFracUs[OSD_STATE_COUNT] = { 0 };
-    static uint16_t osdElementDurationUs[OSD_ITEM_COUNT] = { 0 };
-    static uint16_t osdElementGroupMembership[OSD_ITEM_COUNT];
+    static uint32_t osdElementDurationUs[OSD_ITEM_COUNT] = { 0 };
+    static uint8_t osdElementGroupMembership[OSD_ITEM_COUNT];
     static uint16_t osdElementGroupTargetFracUs[OSD_GROUP_COUNT] = { 0 };
     static uint16_t osdElementGroupDurationFracUs[OSD_GROUP_COUNT] = { 0 };
     static uint8_t osdElementGroup;
